+++ Editor/VisionOSSettingsEditor.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-using RenderMode = UnityEngine.XR.VisionOS.RenderMode;
 
 namespace UnityEditor.XR.VisionOS
 {
@@ -9,40 +8,36 @@ namespace UnityEditor.XR.VisionOS
         const string k_WorldSensingUsageWarning = "World sensing usage description is required if world sensing features " +
             "(images, planes, or meshes) will be used. If this field is blank, the app will not be allowed to request " +
             "world sensing authorization, and these features will not work. If your app does not use world sensing, " +
-            "you can safely ignore this warning.";
-        
+            "you can safely leave this field blank.";
+
         SerializedProperty m_AppModeProperty;
-        SerializedProperty m_DeviceTargetProperty;
-        SerializedProperty m_RenderModeProperty;
-        SerializedProperty m_VolumeModeProperty;
-        SerializedProperty m_VolumeDimensionsProperty;
         SerializedProperty m_HandsTrackingUsageDescriptionProperty;
         SerializedProperty m_WorldSensingUsageDescriptionProperty;
-        
+
         void OnEnable()
         {
             m_AppModeProperty = serializedObject.FindProperty("m_AppMode");
-            m_DeviceTargetProperty = serializedObject.FindProperty("m_DeviceTarget");
-            m_RenderModeProperty = serializedObject.FindProperty("m_RenderMode");
-            m_VolumeModeProperty = serializedObject.FindProperty("m_VolumeMode");
-            m_VolumeDimensionsProperty = serializedObject.FindProperty("m_VolumeDimensions");
             m_HandsTrackingUsageDescriptionProperty = serializedObject.FindProperty("m_HandsTrackingUsageDescription");
             m_WorldSensingUsageDescriptionProperty = serializedObject.FindProperty("m_WorldSensingUsageDescription");
         }
 
         public override void OnInspectorGUI()
         {
+            var isLoaderEnabled = VisionOSBuildProcessor.IsLoaderEnabled();
+
             serializedObject.Update();
             EditorGUIUtility.labelWidth = 200;
-            
+
             EditorGUILayout.LabelField("General Settings", EditorStyles.boldLabel);
 
-            var isLoaderEnabled = VisionOSBuildProcessor.IsLoaderEnabled();
-            using (new EditorGUI.DisabledScope(!isLoaderEnabled))
-            {
-                EditorGUILayout.PropertyField(m_HandsTrackingUsageDescriptionProperty);
+            // TODO: Force app mode to VR if PolySpatial packages are missing
+            EditorGUILayout.PropertyField(m_AppModeProperty);
+            var appMode = (VisionOSSettings.AppMode)m_AppModeProperty.intValue;
 
+            void ShowUsageDescriptionFields()
+            {
                 // TODO: Enable/disable hand tracking
+                EditorGUILayout.PropertyField(m_HandsTrackingUsageDescriptionProperty);
                 if (isLoaderEnabled && string.IsNullOrEmpty(m_HandsTrackingUsageDescriptionProperty.stringValue))
                     EditorGUILayout.HelpBox(VisionOSBuildProcessor.HandTrackingUsageWarning, MessageType.Warning);
 
@@ -50,55 +45,40 @@ namespace UnityEditor.XR.VisionOS
                 if (isLoaderEnabled && string.IsNullOrEmpty(m_WorldSensingUsageDescriptionProperty.stringValue))
                     EditorGUILayout.HelpBox(k_WorldSensingUsageWarning, MessageType.Warning);
             }
-            
-            // TODO: Force app mode to VR if polyspatial packages are missing
-            EditorGUILayout.PropertyField(m_AppModeProperty);
-            
-            EditorGUILayout.Space();
-            
-            EditorGUILayout.LabelField("Virtual Reality Settings", EditorStyles.boldLabel);
 
-            using (new EditorGUI.DisabledScope(m_AppModeProperty.enumValueIndex != (int)VisionOSSettings.AppMode.VR))
+            if (appMode == VisionOSSettings.AppMode.VR)
             {
-                using (var check = new EditorGUI.ChangeCheckScope())
+                EditorGUILayout.Space();
+                EditorGUILayout.LabelField("Virtual Reality settings", EditorStyles.boldLabel);
+
+                ShowUsageDescriptionFields();
+
+#if !(UNITY_2022_3_9 || UNITY_2022_3_10)
+                switch (PlayerSettings.VisionOS.sdkVersion)
                 {
-                    EditorGUILayout.PropertyField(m_DeviceTargetProperty);
-
-                    if (check.changed)
-                    {
-                        var mode = (VisionOSSettings.DeviceTarget)m_DeviceTargetProperty.intValue;
-                        switch (mode)
-                        {
-                            case VisionOSSettings.DeviceTarget.Device:
-                                m_RenderModeProperty.intValue = (int)RenderMode.SinglePassInstanced;
-                                break;
-                            case VisionOSSettings.DeviceTarget.Simulator:
-                                m_RenderModeProperty.intValue = (int)RenderMode.MultiPass;
-                                break;
-                            default:
-                                Debug.LogError($"Unsupported device mode {mode}");
-                                break;
-                        }
-                    }
+                    case VisionOSSdkVersion.Device:
+                        EditorGUILayout.HelpBox("When building for visionOS Device SDK, Single-Pass Instanced rendering will be used.", MessageType.Info);
+                        break;
+                    case VisionOSSdkVersion.Simulator:
+                        EditorGUILayout.HelpBox("When building for visionOS Simulator SDK, Multi-Pass rendering will be used.", MessageType.Info);
+                        break;
                 }
+#endif
             }
-
-            using (new EditorGUI.DisabledScope(true))
+            else if (appMode == VisionOSSettings.AppMode.MR)
             {
-                EditorGUILayout.PropertyField(m_RenderModeProperty);
-            }
-            
-            EditorGUILayout.Space();
-            EditorGUILayout.LabelField("Mixed Reality Settings", EditorStyles.boldLabel);
+                EditorGUILayout.Space();
+                EditorGUILayout.LabelField("Mixed Reality settings", EditorStyles.boldLabel);
 
-            using (new EditorGUI.DisabledScope(m_AppModeProperty.enumValueIndex != (int) VisionOSSettings.AppMode.MR))
-            {
-                EditorGUILayout.PropertyField(m_VolumeModeProperty);
+#if UNITY_HAS_POLYSPATIAL_VISIONOS
+                ShowUsageDescriptionFields();
 
-                using (new EditorGUI.DisabledScope(m_VolumeModeProperty.enumValueIndex == (int)VisionOSSettings.VolumeMode.Unbounded))
-                {
-                    EditorGUILayout.PropertyField(m_VolumeDimensionsProperty);
-                }
+                EditorGUILayout.HelpBox(
+                    "The initial window configuration at app launch is determined by the default volume settings, found in Project Settings > PolySpatial Settings.",
+                    MessageType.Info);
+#else
+                EditorGUILayout.HelpBox("Mixed Reality on visionOS requires PolySpatial and the PolySpatial visionOS packages.", MessageType.Error);
+#endif
             }
 
             serializedObject.ApplyModifiedProperties();